add webapp logging and format code with csharpier

+++ src/Controllers/UserController.cs
@@ -3,15 +3,20 @@ using magnus_backend.Models;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 
-//"Url": "http://192.168.1.29:5000" ADD BACK TO APPSETTINGS
-
 namespace magnus_backend.Controllers;
 
 [Route("api/user")]
 [ApiController]
-public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
+public class UserController : ControllerBase, IUser
 {
-    private readonly IMongoClient _mongoClient = mongoClient;
+    private readonly IMongoClient _mongoClient;
+    private readonly ILog _logger;
+
+    public UserController(IMongoClient mongoClient, ILog logger)
+    {
+        _mongoClient = mongoClient;
+        _logger = logger;
+    }
 
     // Here is where we create all API's that correspond to users
     [HttpGet("{UserId}")]
@@ -26,6 +31,11 @@ public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
 
         if (user.Count == 0)
         {
+            _logger.Log(
+                $"No user found with UserId {UserId}",
+                Enums.LogLevels.Info,
+                "UserController - GetUser"
+            );
             return NotFound(new { Message = $"No user found with UserId {UserId}" });
         }
 
@@ -50,9 +60,9 @@ public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
     [HttpPost]
     public IActionResult AddUser([FromBody] UserModel user)
     {
-
         if (user == null)
         {
+            _logger.Log($"User cannot be null", Enums.LogLevels.Error, "UserController - AddUser");
             return BadRequest("User cannot be null");
         }
 
@@ -68,9 +78,11 @@ public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
         }
 
         // make sure all of our values are filled out
-        if (!string.IsNullOrWhiteSpace(user.UserId) &&
-            !string.IsNullOrWhiteSpace(user.Username) &&
-            !string.IsNullOrWhiteSpace(user.Email))
+        if (
+            !string.IsNullOrWhiteSpace(user.UserId)
+            && !string.IsNullOrWhiteSpace(user.Username)
+            && !string.IsNullOrWhiteSpace(user.Email)
+        )
         {
             try
             {
@@ -78,16 +90,29 @@ public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
                 user.CreatedAt = DateTime.UtcNow.ToString();
 
                 collection.InsertOne(user);
-                Console.WriteLine($"Created user: {user.Username}");
+                _logger.Log(
+                    $"Created User Successfully: {user.UserId}",
+                    Enums.LogLevels.Info,
+                    "UserController - AddUser"
+                );
                 return Ok("Created User Successfully");
             }
             catch (Exception e)
             {
-                Console.WriteLine($"Failed to create user: {e}");
+                _logger.Log(
+                    $"Failed to create user: {e}",
+                    Enums.LogLevels.Error,
+                    "UserController - AddUser"
+                );
                 return BadRequest("Failed to create User");
             }
         }
 
+        _logger.Log(
+            $"User values cannot be null or whitespace!",
+            Enums.LogLevels.Error,
+            "UserController - AddUser"
+        );
         return BadRequest("User values cannot be null or whitespace!");
     }
 
@@ -102,14 +127,36 @@ public class UserController(IMongoClient mongoClient) : ControllerBase, IUser
         var database = _mongoClient.GetDatabase("Magnus");
         var collection = database.GetCollection<UserModel>("users");
 
+        var filter = Builders<UserModel>.Filter.Eq("UserId", userId);
+        var users = collection.Find(filter).ToList();
+
+        if (users.Count == 0)
+        {
+            _logger.Log(
+                $"User does not exist: {userId}",
+                Enums.LogLevels.Info,
+                "UserController - DeleteUser"
+            );
+            return BadRequest($"User does not exist: {userId}");
+        }
+
         try
         {
-            var filter = Builders<UserModel>.Filter.Eq("UserId", userId);
             collection.DeleteOne(filter);
+            _logger.Log(
+                $"Deleted User Successfully. UserId: {userId}",
+                Enums.LogLevels.Info,
+                "UserController - DeleteUser"
+            );
             return Ok($"Deleted User Successfully. UserId: {userId}");
         }
         catch (Exception e)
         {
+            _logger.Log(
+                $"Failed to delete user: {userId}. Error: {e}",
+                Enums.LogLevels.Error,
+                "UserController - DeleteUser"
+            );
             return BadRequest($"Failed to delete User: {e}");
         }
     }
